reorder content and move award slice from about us
