+++ site/pages/about-us/index.js
@@ -4,7 +4,6 @@ import React from 'react';
 import TimelineSlice from './timeline-slice';
 import PrinciplesSlice from './principles-slice';
 import SocialSlice from './social-slice';
-import AwardsSlice from './awards-slice';
 import QAndASlice from './q-and-a-slice';
 import ChecklistContactUs from '../../slices/checklist-contact-us-slice';
 import type { Tweet, InstagramPost } from '../../types/';
@@ -24,7 +23,6 @@ const AboutUs = ({ tweets, instagramPosts, qAndAs }: AboutUsProps) => {
       <ChecklistContactUs />
       <QAndASlice qAndAs={qAndAs} />
       <SocialSlice tweets={tweets} instagramPosts={instagramPosts} />
-      <AwardsSlice />
     </div>
   );
 };
